simplify downloader (remove guessing about how manager would work)

+++ rover/download.py
@@ -19,17 +19,15 @@ TMPEXPIRE = 60 * 60 * 24
 
 class Downloader(SqliteSupport):
     """
-    Download a single request (typically for a day), ingest and
-    index it.
+    Download a single request (typically for a day), ingest and index it.
 
-    The only complex thing here is that these may run in parallel
-    and call ingest.  That means that multiple ingest instances
-    can be running in parallel, all using temp tables in the database.
-    So we need to manage temp table names and clear out old tables
-    from crashed processes.
+    The only complex thing here is that these may run in parallel and call ingest.
+    That means that multiple ingest instances can be running in parallel, all using
+    temp tables in the database.  So we need to manage temp table names and clear
+    out old tables from crashed processes.
 
-    To do this we have another table that is a list of retrievers,
-    along with URLs, PIDs, table names and epochs.
+    To do this we have another table that is a list of retrievers, along with
+    URLs, PIDs, table names and epochs.
 
     This isn't a problem for the main ingest command because only
     a sin=gle instance of the command line command runs at any one
@@ -58,30 +56,23 @@ class Downloader(SqliteSupport):
         """
         self._assert_single_use()
         retrievers_id, table = self._update_retrievers_table(url)
+        path = None
         try:
             path = self._do_download(url)
             self._ingester.ingest([path], table=table)
             self._indexer.index()
-            unlink(path)
         finally:
+            if path: unlink(path)
             self._execute('drop table if exists %s' % table)
             self._execute('delete from rover_downloaders where id = ?', (retrievers_id,))
 
     def _update_retrievers_table(self, url):
         self._clear_dead_retrievers()
-        # either we already have an entry in the table for our url,
-        # in which case url is present and pid is not, or we need
-        # to create an entry ourselves.
-        try:
-            id, pid, table = self._fetchone('select id, pid, table_name from rover_downloaders where url like ?', (url,))
-            if pid and pid != getpid():
-                raise Exception('A retriever already exists for %s' % url)
-            self._execute('update rover_downloaders set pid = ? where id = ?', (getpid(), id))
-        except NoResult:
-            pid = getpid()
-            table = self._retrievers_table_name(url, pid)
-            self._execute('insert into rover_downloaders (pid, table_name, url) values (?, ?, ?)', (pid, table, url))
-            id = self._fetchsingle('select id from rover_downloaders where url like ?', (url,))
+        pid = getpid()
+        table = self._retrievers_table_name(url, pid)
+        self._execute('insert into rover_downloaders (pid, table_name, url) values (?, ?, ?)', (pid, table, url))
+        id = self._fetchsingle('select id from rover_downloaders where pid = ? and table_name like ? and url like ?',
+                               (pid, table, url))  # todo - auto retrieve key?
         return id, table
 
     def _do_download(self, url):
